Move MATERIAL_BIND_GROUP_INDEX, ErasedMaterialPipelineKey, ErasedMaterialKey, ErasedMaterialKeyVTable, RenderPhaseType
